NC references with c. coordinates are now accepted (draft).
- If a gene name is provided as a transcript variant, e.g.,
NC_000001.11(OR4F5_v001):c.101del, we query the NC database to get
all the transcripts part of that gene, as well as those that are in
their immediate suroundings and construct the mutalyzer record from
them.
- If a transcript accession.version is provided as a transcript
variant, e.g, NC_000001.11(NM_001302680.1):c.62825del, we query the
DB to get that transcript as well as those that re in its immediate
vecinity as above.
- Otherwise we construct an almost empty record that contains all
the genes names with empty CDS and transcript lists. This is such
the user can see the list of genes in the error.